refactor: rename DTO folder
